# services/catalog-service/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AV Art Works - Catalog Service
  version: 1.0.0
  description: Products, categories, inventory, search/filter, and image handling.
servers:
  - url: http://localhost:8082
paths:
  /api/v1/products:
    get:
      summary: List products
      parameters:
        - { name: q, in: query, schema: { type: string }, description: full-text search }
        - { name: category, in: query, schema: { type: string }, description: category slug or id }
        - { name: medium, in: query, schema: { type: string, enum: [resin, texture, acrylic, custom, handmade] } }
        - { name: sort, in: query, schema: { type: string, enum: [newest, price_asc, price_desc, title] } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 12 } }
      responses:
        "200": { description: Paginated products }
    post:
      summary: Create a product (admin)
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Created }
        "403": { description: Admin required }
  /api/v1/products/{slug}:
    get:
      summary: Get product detail by slug
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Product }
        "404": { description: Not found }
  /api/v1/products/{id}:
    put:
      summary: Update a product (admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Updated } }
    delete:
      summary: Delete a product (admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "204": { description: Deleted } }
  /api/v1/products/{id}/inventory:
    patch:
      summary: Set product inventory (admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Updated } }
  /api/v1/products/{id}/images:
    post:
      summary: Attach an image to a product (admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "201": { description: Created } }
  /api/v1/uploads/presign:
    post:
      summary: Get a presigned upload URL (admin)
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Presign target } }
  /api/v1/categories:
    get:
      summary: List categories
      responses: { "200": { description: Categories } }
    post:
      summary: Create a category (admin)
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Created } }
  /internal/products/{id}/reserve:
    post:
      summary: Reserve stock (service-to-service)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Reserved }
        "409": { description: Insufficient stock }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT